backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
channels==4.2.2
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
decouple==0.0.7
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.3.1
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.5.0
gunicorn==23.0.0
huggingface-hub==0.30.2
idna==3.10
msgpack==1.1.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
reportlab==4.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tokenizers==0.21.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.4.0
whitenoise==6.8.0
